refactor: Remove cms page
