+++ src/pages/index.tsx
@@ -3,31 +3,31 @@ import 'src/styles/pages/homepage.scss'
 import { useSession } from '@faststore/sdk'
 import { graphql } from 'gatsby'
 import { GatsbySeo, JsonLd } from 'gatsby-plugin-next-seo'
-import { Suspense } from 'react'
-import BannerText from 'src/components/sections/BannerText'
-import Hero from 'src/components/sections/Hero'
-import IncentivesHeader from 'src/components/sections/Incentives/IncentivesHeader'
-import { incentivesMockHeader as IncentivesMock } from 'src/components/sections/Incentives/incentivesMock'
-import ProductShelf from 'src/components/sections/ProductShelf'
-import ProductTiles from 'src/components/sections/ProductTiles'
-import ProductShelfSkeleton from 'src/components/skeletons/ProductShelfSkeleton'
-import ProductTilesSkeleton from 'src/components/skeletons/ProductTilesSkeleton'
-import { ITEMS_PER_SECTION } from 'src/constants'
 import { mark } from 'src/sdk/tests/mark'
 import type { PageProps } from 'gatsby'
-import type { HomePageQueryQuery } from '@generated/graphql'
+import type { CmsHomePageQueryQuery } from '@generated/graphql'
+import RenderPageSections from 'src/components/cms/RenderPageSections'
+import { clientCMS } from 'src/client'
+import type { ContentData } from '@vtex/client-cms'
 
-export type Props = PageProps<HomePageQueryQuery>
+export type Props = PageProps<
+  CmsHomePageQueryQuery,
+  unknown,
+  unknown,
+  ContentData
+>
 
 function Page(props: Props) {
   const {
     data: { site },
+    location: { pathname, host },
+    serverData: cmsHome,
   } = props
 
   const { locale } = useSession()
 
   const title = site?.siteMetadata?.title ?? ''
-  const siteUrl = `${site?.siteMetadata?.siteUrl}`
+  const siteUrl = `https://${host}${pathname}`
 
   return (
     <>
@@ -69,63 +69,40 @@ function Page(props: Props) {
         If needed, wrap your component in a <Section /> component
         (not the HTML tag) before rendering it here.
       */}
-      <Hero
-        title="New Products Available"
-        subtitle="At BaseStore you can shop the best tech of 2022. Enjoy and get 10% off on your first purchase."
-        linkText="See all"
-        link="/technology"
-        imageSrc="https://storeframework.vtexassets.com/arquivos/ids/190897/Photo.jpg"
-        imageAlt="Quest 2 Controller on a table"
-      />
-
-      <IncentivesHeader incentives={IncentivesMock} />
-
-      <Suspense fallback={<ProductShelfSkeleton loading />}>
-        <ProductShelf
-          first={ITEMS_PER_SECTION}
-          selectedFacets={[{ key: 'productClusterIds', value: '140' }]}
-          title="Most Wanted"
-        />
-      </Suspense>
-
-      <Suspense fallback={<ProductTilesSkeleton loading />}>
-        <ProductTiles
-          first={3}
-          selectedFacets={[{ key: 'productClusterIds', value: '141' }]}
-          title="Just Arrived"
-        />
-      </Suspense>
-
-      <BannerText
-        title="The sun has set on our Summer Sale! Save up to 50% off. Don't miss out!"
-        actionPath="/"
-        actionLabel="Call to action"
-        colorVariant="light"
-      />
-
-      <Suspense fallback={<ProductShelfSkeleton loading />}>
-        <ProductShelf
-          first={ITEMS_PER_SECTION}
-          selectedFacets={[{ key: 'productClusterIds', value: '142' }]}
-          title="Deals & Promotions"
-        />
-      </Suspense>
+      <RenderPageSections sections={cmsHome.sections} />
     </>
   )
 }
 
 export const querySSG = graphql`
-  query HomePageQuery {
+  query CmsHomePageQuery {
     site {
       siteMetadata {
         title
         description
         titleTemplate
-        siteUrl
       }
     }
   }
 `
+
+export async function getServerData() {
+  try {
+    const cmsHome = await clientCMS.getCMSPage({
+      contentType: 'home',
+      documentId: 'ad2fd81d-a53c-4281-8d01-a4fc2f274db3',
+      versionId: '1b18fbcf-ce08-4ead-9011-364921e725c3',
+    })
+
+    return {
+      props: cmsHome,
+    }
+  } catch (error) {
+    return {
+      status: 500,
+    }
+  }
+}
 
 Page.displayName = 'Page'
 export default mark(Page)
