+++ store.config.js
@@ -36,7 +36,6 @@ module.exports = {
       home: '/',
       pdp: '/apple-magic-mouse/p',
       collection: '/office',
-      cmsHome: '/cms',
     },
   },
 
@@ -49,7 +48,6 @@ module.exports = {
       collection_filtered:
         '/office/?category-1=office&marca=acer&facets=category-1%2Cmarca',
       search: '/s?q=orange',
-      cmsHome: '/cms',
     },
   },
